+++ test/unit/test_arguments.py
@@ -60,7 +60,9 @@ def test_io_args(test_params, expected_params):
 
 
 @pytest.mark.parametrize("test_params, expected_params", [
-    ('', dict(quiet=False, loglevel='INFO')),
+    ('', dict(quiet=False,
+              loglevel='INFO',
+              no_logfile=False)),
 ])
 def test_logging_args(test_params, expected_params):
     _test_args(test_params, expected_params, arguments.add_logging_args)